[fix] fixed behaviour of deployment if configservice url can not be found

+++ opsicommon/deployment/posix.py
@@ -119,7 +119,10 @@ class PosixDeployThread(DeployThread):
 				logger.debug("Will use: %s", opsiscript)
 				self._executeViaSSH(f"chmod +x {opsiscript}")
 
-				service_address = self.backend.configState_getObjects(configId='clientconfig.configserver.url', objectId=hostObj.id)[0].values[0]
+				service_configstate = self.backend.configState_getObjects(configId='clientconfig.configserver.url', objectId=hostObj.id)
+				if len(service_configstate) != 1 and len(service_configstate[0].values) != 1:
+					raise ValueError("Could not determine associated configservice url")
+				service_address = service_configstate[0].values[0]
 				finalize = "noreboot"
 				if self.reboot:
 					finalize = "reboot"
