+++ opsicommon/deployment/windows.py
@@ -70,7 +70,10 @@ class WindowsDeployThread(DeployThread):
 			self._installWithServersideMount()
 
 	def install_from_path(self, path, hostObj):
-		service_address = self.backend.configState_getObjects(configId='clientconfig.configserver.url', objectId=hostObj.id)[0].values[0]
+		service_configstate = self.backend.configState_getObjects(configId='clientconfig.configserver.url', objectId=hostObj.id)
+		if len(service_configstate) != 1 and len(service_configstate[0].values) != 1:
+			raise ValueError("Could not determine associated configservice url")
+		service_address = service_configstate[0].values[0]
 		finalize = "noreboot"
 		if self.reboot:
 			finalize = "reboot"
